moving common type parsers and formatters to utilities
